Refactor shell into handlers and ndb modules
